// frontend/booking-vehiculos/src/app/core/api/model/reservaRequestDto.ts
/**
 * Microservicio.Coche.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ReservaRequestDto { 
    VehiculoId?: string;
    ClienteId?: string;
    SucursalRetiroId?: string;
    SucursalEntregaId?: string;
    FechaRetiro?: string;
    FechaEntrega?: string;
}