// frontend/booking-vehiculos/src/app/core/api/model/crearPagoRequest.ts
/**
 * Microservicio.Coche.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CrearPagoRequest { 
    RES_id?: string;
    PAG_monto?: number;
    PAG_metodo?: string | null;
    PAG_estado?: string | null;
}